Auth0 authentication: Fetch user profile from Auth0 only when creating a new user
